backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Anton <patrin_toxa@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
drf-spectacular = "^0.27.2"
django-cors-headers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
